danielle/requirements.txt
asgiref==3.11.1
attrs==26.1.0
certifi==2026.2.25
charset-normalizer==3.4.6
colorama==0.4.6
coverage==7.4.4
dj-database-url==3.1.2
Django==4.2.11
django-cors-headers==4.3.1
django-filter==24.1
django-heroku==0.3.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
idna==3.11
inflection==0.5.1
iniconfig==2.3.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
mysqlclient==2.2.8
packaging==26.0
pluggy==1.6.0
psycopg2==2.9.11
Pygments==2.19.2
PyJWT==2.12.1
pytest==9.0.2
pytest-django==4.8.0
python-dotenv==1.2.2
PyYAML==6.0.3
referencing==0.37.0
requests==2.31.0
rpds-py==0.30.0
sqlparse==0.5.5
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.3
validate-docbr==1.10.0
whitenoise==6.12.0
